Predicting from Emax regression models

Predicting from Emax regression models

Usage

# S3 method for class 'emaxnls'
predict(object, newdata = NULL, interval = "none", level = 0.95, ...)

Arguments

object

An emaxnls object

newdata

A named list or data frame in which to look for variables with which to predict. If newdata is missing the fitted values at the original data points are returned.

interval

A character string indicating if prediction intervals or a confidence interval on the mean responses are to be calculated. Can be "none", "confidence", or "prediction"

level

A numeric scalar between 0 and 1 giving the confidence level for the intervals (if any) to be calculated.

...

Ignored

Value

The predict() method for Emax regression objects always returns a data frame or tibble

Details

Note that predict() for Emax regression calls investr::predFit() to obtain confidence and prediction intervals. Please see the documentation for that function.